Fix bai 3
